refactor(qa-lab): split suite execution
